Tested some alternative implementations.
I feel like the biggest problem right now is not getting the code to do
what I want, but getting it to run fast enough. I've replaced some lines
with lines I've determined to do the same things faster and I've created
a getWordsTest and getTargetsTest to represent more sophisticated
versions of getWords. What I really need right now is to start throwing
some of this python code down to C and try to eke out performance gains
that way.